+++ src/pages/attributes/form/attributesFormFields.tsx
@@ -8,8 +8,11 @@ import {
     StandardFormField,
     StandardFormSectionTitle,
     StandardFormSectionDescription,
-    DefaultIdentifiableFields,
     DescriptionField,
+    NameField,
+    ShortNameField,
+    CodeField,
+    ValueTypeField,
 } from '../../../components'
 import {
     useSchemaSectionHandleOrThrow,
@@ -18,7 +21,6 @@ import {
 import { AttributeTypeComponent } from './AttributeTypeComponent'
 import { OptionSetField } from './OptionSetField'
 import { SortOrderField } from './SortOrderField'
-import { ValueTypeField } from './ValueTypeField'
 
 export const AttributeFormFields = ({
     initialValues,
@@ -27,6 +29,7 @@ export const AttributeFormFields = ({
 }) => {
     useSchemaSectionHandleOrThrow()
     useSyncSelectedSectionWithScroll()
+    const schemaSection = useSchemaSectionHandleOrThrow()
 
     return (
         <SectionedFormSections>
@@ -37,7 +40,20 @@ export const AttributeFormFields = ({
                 <StandardFormSectionDescription>
                     {i18n.t('Set up the basic information for this attribute.')}
                 </StandardFormSectionDescription>
-                <DefaultIdentifiableFields shortNameIsRequired={false} />
+                <StandardFormField>
+                    <NameField schemaSection={schemaSection} />
+                </StandardFormField>
+
+                <StandardFormField>
+                    <ShortNameField
+                        schemaSection={schemaSection}
+                        isRequired={false}
+                    />
+                </StandardFormField>
+
+                <StandardFormField>
+                    <CodeField schemaSection={schemaSection} />
+                </StandardFormField>
                 <StandardFormField>
                     <DescriptionField
                         helpText={i18n.t(