+++ src/pages/programsWip/form/formDescriptor.tsx
@@ -7,9 +7,21 @@ export const ProgramFormDescriptor = {
     label: 'Program',
     sections: [
         {
-            name: 'setup',
+            name: 'enrollmentDetails',
             label: i18n.t('Program Details'),
-            fields: [{ name: 'name', label: i18n.t('Name') }],
+            fields: [
+                { name: 'name', label: i18n.t('Name') },
+                { name: 'shortName', label: i18n.t('Short name') },
+                { name: 'code', label: i18n.t('Code') },
+                { name: 'description', label: i18n.t('Description') },
+                { name: 'version', label: i18n.t('Version') },
+                { name: 'featureType', label: i18n.t('Feature type') },
+                { name: 'relatedProgram', label: i18n.t('Related program') },
+                {
+                    name: 'categoryCombo',
+                    label: i18n.t('Category combination'),
+                },
+            ],
         },
         {
             name: 'enrollmentSettings',
@@ -18,6 +30,28 @@ export const ProgramFormDescriptor = {
                 {
                     name: 'trackedEntityType',
                     label: i18n.t('Tracked entity type'),
+                },
+                {
+                    name: 'onlyEnrollOnce',
+                    label: i18n.t('Only allow one enrollment'),
+                },
+                {
+                    name: 'selectEnrollmentDatesInFuture',
+                    label: i18n.t('Allow enrollment dates in the future'),
+                },
+                {
+                    name: 'displayIncidentDate',
+                    label: i18n.t('Collect an incident date'),
+                },
+                {
+                    name: 'selectIncidentDatesInFuture',
+                    label: i18n.t('Allow incident dates in the future'),
+                },
+                {
+                    name: 'useFirstStageDuringRegistration',
+                    label: i18n.t(
+                        'Collect data for the first stage during enrollment'
+                    ),
                 },
             ],
         },