+++ src/pages/trackedEntityTypes/fields/index.ts
@@ -1,0 +1,5 @@
+export { AttributesTransferField } from './AttributesTransferField'
+export { AttributesConfigurationField } from './AttributesConfigurationField'
+export { AllowAuditLogField } from './AllowAuditLogField'
+export { MinAttributesRequiredField } from './MinAttributesRequiredField'
+export { MaxTeiCountField } from './MaxTeiCountField'